ref draft-js-plugins/draft-js-plugins#903 add support for email and mailto: links
plugin now try to normaliseMail (add 'mailto:' to a given url), and test it against a mailRegex to see if it is a valid email address (with mailto: prepended).

if link is a valid email address, it will be accepted as a mailto:email@domain link and the existing URL test logic will be bypassed.

if link is not a valid email address, it will fall back to existing URL test logic.