refactor: update Entry focus state strategy
This removes previous commit d0982cc and follows the existing pattern
for checking whether or not an Entry is focused to apply the focus
style. This approach prevents a breaking change and still uses the
emojiSelectPopoverEntryFocused theme class.

There is an opportunity here for a future performance refactor, using
native css hover, focus, active selectors rather than internal component
state focus tracking.

This also changes the build number from 4.7.0 to 4.6.6 as this is now a
patch version